requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2.1
django-cors-headers==3.0.1
django-user-agents==0.3.2
djangorestframework==3.9.4
idna==2.8
numpy==1.16.3
oauthlib==3.0.1
pandas==0.24.2
pip==19.1.1
Pygments==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
requests==2.21.0
requests-oauthlib==1.2.0
setuptools==41.0.1
six==1.12.0
sqlparse==0.3.0
ua-parser==0.8.0
urllib3==1.24.3
user-agents==2.0
wheel==0.33.4
